Buscar de usuario por id na base

+++ src/main/kotlin/com/reis/usuario/repository/UserRepository.kt
@@ -3,7 +3,8 @@ package com.reis.usuario.repository
 import com.reis.usuario.model.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
+import java.math.BigInteger
 
 @Repository
-interface UserRepository : JpaRepository<User, Int> {
+interface UserRepository : JpaRepository<User, BigInteger> {
 }
